Update STAT345_final_monopoly.Rmd
Added functions to handle chance cards; have NOT factored this into our movement functions yet. All the functions do now are read you the text of the action card and move you to a new location if movement is necessary. (Movement is appearently all we need to worry about if I am reading the directions correctly.) 

Also changed the text of the roll_to_move_player function to show when we roll doubles twice.